added projectionengine, bubbletracker classes, but only wrote headers

+++ Bubbles/Bubbles/BubbleState.h
@@ -10,7 +10,7 @@ class IBubbleState{
 	public:
 		virtual void addBubble(Bubble b)=0;
 		virtual unsigned int getUnknownBubble(void)=0;
-		virtual void updateBubble(unsigned int ID, int x, int y, int z)=0;
+		virtual void updateBubble(unsigned int ID, float x, float y, float z)=0;
 		virtual map<unsigned int, Bubble> getCurrentState(void)=0;
 		virtual void notifyUserBreaks(unsigned int ID)=0;
 		virtual void notifyTrackingFrame(void)=0;
